+++ CHANGELOG.rst
@@ -9,6 +9,12 @@ Package maintainers and users who have to manually update their installation
 may want to subscribe to `GitHub's tag feed
 <https://github.com/pimutils/vdirsyncer/tags.atom>`_.
 
+Version 0.16-dev
+==============
+
+- Add post_hook for delete event and metadata
+- Add a second argument of post_hook : which action triggered the hook
+
 Version 0.16.7
 ==============